Return number of Draft and Queued Posts for a blog
Tumblr returns the total number of Draft Posts and the total number Queued posts for a blog, but they were not accessible through Jumblr. This requires an app to step through all the Drafts and Queued posts to "count" them; very much a waste of resources. Since Tumblr returns these values already, these changes allow the application to access the counts without having to "count them".