Upgrade com.scribe to 1.3.7
Use the final 1.x release of Scribe. Scribe 1.3.5 adds a new method
OAuthRequest#setFollowRedirects(), which is needed to be able to
(easily) set no-follow on a single request, instead of globally. There
were no concerning changes between 1.3.5 and 1.3.7, so we might as well
bump to the most recent non-breaking version.